+++ src/gcf/__init__.py
@@ -38,7 +38,7 @@ from .constants import KIND_ABBREV, KIND_EXPAND
 from .decode import DecodeError, decode
 from .delta import encode_delta
 from .encode import encode
-from .generic import encode_generic
+from .generic import encode_generic, GenericOptions
 from .session import Session, encode_with_session
 from .decode_generic import decode_generic
 from .stream import StreamEncoder
@@ -62,6 +62,7 @@ __all__ = [
     "encode",
     "encode_delta",
     "encode_generic",
+    "GenericOptions",
     "encode_with_session",
 ]